Phase 3d.3b: consolidate stage-3 orchestration into shared _strip_and_orchestrate
Both strip_traits_for_validation() (schemas.py, for config_flow validation)
and _strip_schema_extensions() (coordinator.py, for gateway feeding) now
call the same shared _strip_and_orchestrate() function. This ensures
validation-passing schemas match what the gateway actually receives.

Consolidated logic (previously duplicated & drifted):
- disabled/skipped/foreign-owner filtering (coordinator only, now shared)
- HGI (18:) dropping (coordinator only, now shared)
- None value stripping (coordinator only, now shared)
- device_comments/owner extension key removal (coordinator only, now shared)
- placed_in_lists check (schemas.py only, now shared — fixes coordinator
  bug where a device in remotes[] could also appear in orphans_hvac)
- TCS orphan safety net (coordinator only, now shared)
- empty device entry → orphans (coordinator only, now shared)
- un-disabled trait-only → orphans (coordinator only, now shared)

Also unified:
- _HEAT_PREFIXES: single definition in schemas.py (was 2 identical copies)
- _TCS_ORPHAN_PREFIXES: single definition in schemas.py
- _SCHEMA_EXTENSION_KEYS: single definition in schemas.py
- _DEVICE_ID_RE: schemas.py now uses hex regex (was decimal-only — missed
  hex device IDs like 1F:000002)
- Removed _strip_traits_rf import from coordinator (no longer used directly)

Updated test_strips_faked_bound_scheme_traits to expect placed_in_lists
behavior (device in remotes[] not duplicated in orphans_hvac).
